backend/requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
sentence-transformers>=2.2.0
nltk
weaviate-client
numpy
